refactor. separating components into their own files
